ui fixes improvements
