+++ apps/web/components/GetTickets.tsx
@@ -4,9 +4,10 @@ import { useState } from "react";
 import { Button } from "./ui/Button";
 import { GasFeeCard } from "./GasFeeCard";
 import MetaMaskProvider from "@/providers/MetamaskProvider";
+import { formatEther } from "ethers";
 
 interface GetTicketsProps {
-  ticketPrice: number;
+  ticketPrice: bigint;
 }
 
 const GetTickets: React.FC<GetTicketsProps> = ({ ticketPrice }) => {
@@ -25,7 +26,7 @@ const GetTickets: React.FC<GetTicketsProps> = ({ ticketPrice }) => {
       <div className="p-8 bg-white shadow-xl rounded-xl full md:w-3/4 md:sticky md:top-4 h-fit">
         <div className="flex items-center justify-between mb-4">
           <span className="text-gray-600">Ticket Price</span>
-          <span>{ticketPrice}ETH</span>
+          <span>{formatEther(ticketPrice)}ETH</span>
         </div>
         <div className="flex items-center justify-between mb-4">
           <span className="text-gray-600">Number of tickets</span>
@@ -48,7 +49,12 @@ const GetTickets: React.FC<GetTicketsProps> = ({ ticketPrice }) => {
         </div>
         <div className="flex items-center justify-between mb-4">
           <span className="text-gray-600">Total</span>
-          <span>{(numberOfTickets * ticketPrice).toFixed(2)}ETH</span>
+          <span>
+            {formatEther(
+              (BigInt(numberOfTickets.toString()) * ticketPrice).toString()
+            )}
+            ETH
+          </span>
         </div>
         <div className="flex items-center justify-between mb-4">
           <span className="text-gray-600">Gas fee</span>
