+++ apps/web/components/Navbar.tsx
@@ -3,7 +3,7 @@
 import Link from "next/link";
 import EventSeaLogo from "../public/icons/EventSeaLogo";
 import WalletIcon from "../public/icons/WalletIcon";
-import CreateEvent from "@/components/create-event-form";
+import CreateEvent from "@/components/create-event/create-event-form";
 import { Button } from "./ui/Button";
 import { SearchBar } from "./SearchBar";
 import { useState } from "react";
@@ -71,21 +71,21 @@ export const ConnectWalletButton = () => {
 export const NavBar = () => {
   return (
     <nav className="flex items-center justify-between max-w-screen-xl px-6 mx-auto bg-white border py-7 rounded-xl">
-      <Link href="/" className="flex gap-1 px-6">
-        <EventSeaLogo />
-        <span className="hidden text-2xl font-bold sm:block">
-          <span className="text-[#0C200A]">Event</span>
-          <span className="text-[#4C6D07]">Sea</span>
-        </span>
-      </Link>
-      <div className="flex items-center gap-4 px-6">
-        <SearchBar />
-        <CreateEvent />
+      <MetaMaskProvider>
+        <Link href="/" className="flex gap-1 px-6">
+          <EventSeaLogo />
+          <span className="hidden text-2xl font-bold sm:block">
+            <span className="text-[#0C200A]">Event</span>
+            <span className="text-[#4C6D07]">Sea</span>
+          </span>
+        </Link>
+        <div className="flex items-center gap-4 px-6">
+          <SearchBar />
+          <CreateEvent />
 
-        <MetaMaskProvider>
           <ConnectWalletButton />
-        </MetaMaskProvider>
-      </div>
+        </div>
+      </MetaMaskProvider>
     </nav>
   );
 };
