+++ apps/web/package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@heroicons/react": "^2.0.18",
     "@hookform/resolvers": "^3.3.1",
     "@metamask/sdk-react": "^0.7.0",
     "@radix-ui/react-dialog": "^1.0.5",
